Update TemaRestController & AutorRestController
All methods now return ResponseEntity objects

+++ Ejemplo/src/main/java/es/daw/bibliografia/api/AutorRestController.java
@@ -1,7 +1,12 @@
 package es.daw.bibliografia.api;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -22,29 +27,71 @@ public class AutorRestController {
 	@Autowired
 	private CitaService citaService;
 	
-	@PostMapping("/api/autor/guardado")
-	public String addAutor(Autor autor) {
+	
+	@GetMapping("/api/autor")
+	public ResponseEntity<Autor> accessAutor(@RequestParam String nombre) {
+		Optional<Autor> autor = autorService.findOneByNombre(nombre);
+		
+		if (autor.isPresent())
+			return new ResponseEntity<Autor>(autor.get(), HttpStatus.OK);
+		else
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+	}
+	
+	@DeleteMapping("/api/autor")
+	public ResponseEntity<Autor> deleteAutor(@RequestParam String nombre) {
+		Optional<Autor> autor = autorService.findOneByNombre(nombre);
+		
+		if (autor.isPresent()) {
+			autorService.deleteByNombre(nombre);
+			
+			return new ResponseEntity<Autor>(autor.get(), HttpStatus.OK);
+		}
+		else
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+	}
+	
+	
+	@PostMapping("/api/autor")
+	public ResponseEntity<Autor> createAutor(@RequestParam String nombre) {
+		Autor autor = new Autor();
+		autor.setNombre(nombre);
 		autorService.save(autor);
+		
+		Optional<Autor> autorOpt = autorService.findOneByNombre(nombre);
+		
+		if (autorOpt.isPresent()) {
+			autorService.deleteByNombre(nombre);
+			
+			return new ResponseEntity<Autor>(autorOpt.get(), HttpStatus.CREATED);
+		} else {
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 
-
-		return "redirect:/autorshow/".concat(autor.getNombre());
 	}
 	
-	@DeleteMapping("/api/autor/borrar")
-	public Autor deleteTema(@RequestParam String nombre) {
-		
-		Autor deletedAutor= autorService.findOneByNombre(nombre).get();
-		autorService.delete(autorService.findOneByNombre(nombre).get().getId());
-		
-		return deletedAutor;
-	}
+//	@DeleteMapping("/api/autor/borrar")
+//	public Autor deleteTema(@RequestParam String nombre) {
+//		
+//		Autor deletedAutor= autorService.findOneByNombre(nombre).get();
+//		autorService.delete(autorService.findOneByNombre(nombre).get().getId());
+//		
+//		return deletedAutor;
+//	}
 	
-	@DeleteMapping("/api/autor/borrarCita/{id}")
-	public Cita deleteCita(@PathVariable long id) {
+	@DeleteMapping("/api/autor/cita")
+	public ResponseEntity<Cita> deleteAutor(@RequestParam Long id) {
+		Optional<Cita> deletedCita = citaService.findOne(id);
+		
+		if (deletedCita.isPresent()) {
+			citaService.delete(id);
+			
+			return new ResponseEntity<Cita>(deletedCita.get(), HttpStatus.OK);
+		}
+		else
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
-		Cita deletedCita= citaService.findOne(id).get();
-		citaService.delete(id);
-
-		return deletedCita;
 	}
 }
